+++ CHANGELOG.md
@@ -11,3 +11,17 @@ IMPROVEMENTS:
 FIXES:
 
 * minio testing server shutdown fixed
+
+# 0.0.3
+
+IMPROVEMENTS:
+
+* module dependencies upgrade
+
+* Go 1.18 now mandatory
+
+* OpenTelemetry API breaking changes
+
+* Logger updated to support fixed fields appended to every output
+
+* Logger supports debug automatically generated and added to output